SADRNet-main/requirements.txt
h5py==2.10.0
matplotlib==3.1.1  
Cython==0.29.13
numba==0.45.1
numpy==1.16.0   
opencv-python==4.1.1.26
torch==1.2.0
torchvision==0.4.0
Pillow==6.1.0                 
pyrender==0.1.33                
scikit-image==0.15.0                
scipy==1.3.1
tensorflow==1.12.0
keras==2.1.6
